Move record initialization to a shared location
